+++ AWSS3/AWSS3Model.h
@@ -75,6 +75,7 @@ typedef NS_ENUM(NSInteger, AWSS3BucketLocationConstraint) {
     AWSS3BucketLocationConstraintUSGovEast1,
     AWSS3BucketLocationConstraintEUNorth1,
     AWSS3BucketLocationConstraintAPEast1,
+    AWSS3BucketLocationConstraintMESouth1
 };
 
 typedef NS_ENUM(NSInteger, AWSS3BucketLogsPermission) {