+++ AWSS3/AWSS3Resources.m
@@ -1298,7 +1298,8 @@
         \"ca-central-1\",\
         \"us-gov-east-1\",\
         \"eu-north-1\",\
-        \"ap-east-1\"\
+        \"ap-east-1\",\
+        \"me-south-1\"\
       ]\
     },\
     \"BucketLoggingStatus\":{\